refactor: split out subrouters
