+++ worker/sync/src/routers/integration.ts
@@ -1,8 +1,8 @@
 import { IRequest, AutoRouter } from "itty-router";
-import { Env } from "./types";
-import { response } from "./utils/request";
-import { getInstance, getInvitation, getUser } from "./utils/data";
-import { importKey, KEY_PREFIX, verifyKeySignature } from "./utils/crypto";
+import { Env } from "../types";
+import { response } from "../utils/request";
+import { getInstance, getInvitation, getUser } from "../utils/data";
+import { importKey, KEY_PREFIX, verifyKeySignature } from "../utils/crypto";
 import { Invitation, ShareInstanceMeta, User } from "@referee-fyi/share";
 
 const verifyBearerToken = async (request: IRequest, env: Env) => {
